splitting user controller
