+++ shoebox/app/views/admin/admin.scala.html
@@ -37,7 +37,7 @@
 		    <ul class="nav">
 		      <li><a href="@com.keepit.controllers.admin.routes.AdminDashboardController.index">Dashboard</a>
 		      <li><a href="@com.keepit.controllers.admin.routes.AdminBookmarksController.bookmarksView(0)">Keeps</a>
-		      <li><a href="@com.keepit.controllers.routes.UserController.usersView">Users</a>
+		      <li><a href="@com.keepit.controllers.admin.routes.AdminUserController.usersView">Users</a>
 		      <li><a href="@com.keepit.controllers.routes.SocialUserController.socialUsersView(0)">Social Users</a>
 		      <li><a href="@com.keepit.controllers.routes.CommentController.followsView">Follows</a>
 		      <li><a href="@com.keepit.controllers.routes.CommentController.commentsView(0)">Comments</a>
