+++ shoebox/app/views/admin/users.scala.html
@@ -1,4 +1,4 @@
-@(userStatisticsList: Seq[com.keepit.controllers.UserStatistics])
+@(userStatisticsList: Seq[com.keepit.controllers.admin.UserStatistics])
 
 @admin("All " + userStatisticsList.size + " Users") {
   <table class="table table-bordered">
